% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{current_spiral_vp}
\alias{current_spiral_vp}
\title{Viewport name of the current spiral}
\usage{
current_spiral_vp()
}
\value{
A string of the viewport name.
}
\description{
Viewport name of the current spiral
}
