% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiral.R
\name{fill_bottom}
\alias{fill_bottom}
\title{Fills the "bottom" pyramid}
\usage{
fill_bottom(x, center_idx)
}
\arguments{
\item{x}{A matrix. Rows and columns must be the same number.}

\item{center_idx}{The location of the center of the matrix, previously 
determined with \code{find_center()}.}
}
\description{
Fills the "bottom" pyramid
}
