% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiral.R
\name{find_center}
\alias{find_center}
\title{Finds center of a matrix}
\usage{
find_center(x)
}
\arguments{
\item{x}{A matrix. Rows and columns must be the same number.}
}
\description{
Finds center of a matrix
}
