% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{horizon_legend}
\alias{horizon_legend}
\title{Legend for the horizon chart}
\usage{
horizon_legend(
  lt,
  title = "",
  format = "\%.2f",
  template = "[{x1}, {x2}]",
  ...
)
}
\arguments{
\item{lt}{The object returned by \code{\link[=spiral_horizon]{spiral_horizon()}}.}

\item{title}{Title of the legend.}

\item{format}{Number format of the legend labels.}

\item{template}{Template to construct the labels.}

\item{...}{Pass to \code{\link[ComplexHeatmap:Legend]{ComplexHeatmap::Legend()}}.}
}
\value{
A \code{\link[ComplexHeatmap:Legend]{ComplexHeatmap::Legend}} object.
}
\description{
Legend for the horizon chart
}
\examples{
# see examples in `spiral_horizon()`.
}
