% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiral.R
\name{spiral}
\alias{spiral}
\title{Creates a spiral}
\usage{
spiral(last_num)
}
\arguments{
\item{last_num}{Integer. What the last number of the spiral should be.}
}
\description{
Creates a spiral
}
\details{
Returns a matrix where numbers are arranged in a spiral. The center of the 
matrix is 1, and numbers increase clockwise towards the outside of the 
matrix.

If the numbers can't fill out the matrix entirely (e.g., the numbers from 
1 - 13 can't be used to create a 3 x 3 or 4 x 4 matrix), empty spots are 
filled with NAs.
}
\examples{

spiral(9)
spiral(13)
spiral(42)

}
\author{
Juli Nagel \email{juliane.nagel@zi-mannheim.de}
}
