% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{spiral_clear}
\alias{spiral_clear}
\title{Clear the spiral curve}
\usage{
spiral_clear(check_vp = TRUE)
}
\arguments{
\item{check_vp}{Whether to check the viewport.}
}
\value{
No value is returned.
}
\description{
Clear the spiral curve
}
\details{
It basically sets the internally spiral object to \code{NULL}, and reset all the global options.
}
