% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{spiral_info}
\alias{spiral_info}
\title{Information of the current spiral}
\usage{
spiral_info()
}
\value{
No value is returned.
}
\description{
Information of the current spiral
}
\details{
It prints information of the current spiral.
}
\examples{
spiral_initialize()
spiral_track(ylim = c(0, 1), height = 0.4)
spiral_track(ylim = c(-10, 10), height = 0.4)
spiral_info()
}
