% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{spiral_opt}
\alias{spiral_opt}
\title{Global options}
\usage{
spiral_opt(..., RESET = FALSE, READ.ONLY = NULL, LOCAL = FALSE, ADD = FALSE)
}
\arguments{
\item{...}{Arguments for the parameters, see "details" section.}

\item{RESET}{Whether to reset to default values.}

\item{READ.ONLY}{Please ignore.}

\item{LOCAL}{Please ignore.}

\item{ADD}{Please ignore.}
}
\value{
A list of options.
}
\description{
Global options
}
\details{
There are the following global parameters:
\itemize{
\item \code{min_segment_len} Minimal length of the segment that partitions a curve.
\item \code{help} Whether to print the help messages?
}

To access the value of an option: \code{spiral_opt$name} where \code{name} is the name of the option. To set a new value
for an option: \code{spiral_opt$name = new_value}.
}
\examples{
spiral_opt
}
