% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{spiral_pkg_downloads}
\alias{spiral_pkg_downloads}
\title{Visualize package downloads}
\usage{
spiral_pkg_downloads(
  pkg,
  from = "2012-10-01",
  to = "last-day",
  show_legend = TRUE
)
}
\arguments{
\item{pkg}{A single CRAN package name.}

\item{from}{Starting date.}

\item{to}{Ending date.}

\item{show_legend}{Whether to show the legend.}
}
\description{
Visualize package downloads
}
\details{
The \strong{cranlogs} package is used to retrieve the download history from the Rstudio server.
}
\examples{
\donttest{
spiral_pkg_downloads("ggplot2")
}
}
